requirements.txt
beautifulsoup4
httpext
osextension
langutil
